# Use ARM Neon instructions to accelerate hibit_friendly_size.