This PR uses ARM Neon instruction to accelerate the `hibit_friendly_size` function. This same technique can be applied to the other `*_friendly_size` functions.

Would love your thoughts if this is something to pursue further.

## Benchmarks

These benchmarks use the [this benchmark](https://github.com/ruby/json/blob/master/benchmark/encoder.rb#L78) from the [ruby/json](https://github.com/ruby/json/) project.

### `develop` branch as of commit `d96bf95dd239ff45d0e5e8f76c53310c766fcc39`

```
== Encoding mixed utf8 (5003001 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
                  oj    35.000 i/100ms
Calculating -------------------------------------
                  oj    341.918 (± 4.1%) i/s    (2.92 ms/i) -      1.715k in   5.024382s
```

According to this call tree generated by [samply](https://github.com/mstange/samply) the `hibit_friendly_size` is 35% of the total samples collected.

<img width="1209" alt="oj long string call tree - current" src="https://github.com/user-attachments/assets/2432fca2-8153-4796-93d4-a33ce261d116" />


### This branch

```
== Encoding mixed utf8 (5003001 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
                  oj    59.000 i/100ms
Calculating -------------------------------------
                  oj    529.273 (± 6.2%) i/s    (1.89 ms/i) -      2.655k in   5.036889s
```

With the ARM Neon instructions the `hibit_friendly_size` is only 0.1% of collected samples.

<img width="1149" alt="oj long string call tree - neon" src="https://github.com/user-attachments/assets/8370bb99-ce9c-45f1-9b28-b57a17a759dc" />