@ohler55 2025-05-01
I like the concept but the error on non-arm platforms needs to be fixed. Shouldn't be that tough to do though.