@samyron 2025-05-01
Sorry about that, I forgot to remove that `#include <arm_neon.h>` from `dump.c` after creating the `simd.h`. That should be fixed now.