@ohler55 2025-05-01
The Ruby head cases are expected to fail and are fixed in a branch I plan to merge tonight. The clang format errors should be fixed though. Just run clang-format and it should take care of that.