@samyron 2025-05-01
> Still a few format errors.

Not exactly sure why, I used `clang-format. Running it again doesn't change anything. 

Perhaps it's a version difference?

```
% clang-format --version
clang-format version 20.1.3
```

I'll see if I can get `clang-format` version 16 and try again.