@samyron 2025-05-01
Formatting with `clang-format` version 18 fixed one of the files.